+++ src/integration/react.ts
@@ -1,6 +1,7 @@
 import { PureComponent, ReactNode } from 'react';
 
 import type { Persistor } from '../types';
+import { logger } from '../utils';
 
 type Props = {
   onBeforeLift?: () => void | Promise<void>;
@@ -51,12 +52,8 @@ export class PersistGate extends PureComponent<Props, State> {
   }
 
   render(): ReactNode {
-    if (process.env.NODE_ENV !== 'production') {
-      if (typeof this.props.children === 'function' && this.props.loading) {
-        console.error(
-          'redux-persist: PersistGate expects either a function child or loading prop, but not both. The loading prop will be ignored.',
-        );
-      }
+    if (typeof this.props.children === 'function' && this.props.loading) {
+      logger.error('PersistGate expects either a function child or loading prop, but not both. The loading prop will be ignored.');
     }
 
     if (typeof this.props.children === 'function') {